Replace hooks.server.ts with +layout.server.ts
